Added updates to rbd, prep work for Jewel

+++ recipes/maint_osd_remove_zap.rb
@@ -78,6 +78,7 @@ if node['ceph']['osd']['remove']['devices']
 
     execute "remove mount directory - #{osd_device['device']}" do
       command lazy { "rm -rf /var/lib/ceph/osd/#{node['ceph']['cluster']}-#{osd_device['osd']}" }
+      only_if "test -d /var/lib/ceph/osd/#{node['ceph']['cluster']}-#{osd_device['osd']}"
     end
 
     # No guard on this - zap!! The only psuedo guard is if the 'zap' attribute was missing.
